admin/data/loader/requirements.txt
flake8==4.0 
psycopg2-binary==2.9.5
pytest==7.2.0
